backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
contourpy==1.3.3
cryptography==44.0.3
cryptograpy==0.0.0
cufflinks==0.17.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.7
django-admin-interface==0.30.0
django-allauth==65.8.0
django-colorfield==0.14.0
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@d269874c96fe5ff6d231c3017cdd4650bc2f3105
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.33.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.7.0
geographiclib==2.1
geopy==2.4.1
greenlet==3.1.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.3
msgspec==0.19.0
multidict==6.6.4
narwhals==2.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.1
openai==1.78.0
opencage==3.2.0
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.8
plotly==6.3.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-engineio==4.11.2
python-json-logger==3.3.0
python-slugify==8.0.4
python-socketio==5.12.1
pytz==2025.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
redis==4.6.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
scipy==1.16.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.20
uuid==1.30
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wsproto==1.2.0
yarl==1.20.1
